+++ utils/cdk-docker-execute.sh
@@ -22,43 +22,80 @@ bumprc() { # return the higher absolute value of the inputs
 
 RC=0
 
+#
+# This is used to allow/accept files which have spaces in their names
+#
 IFS=$'\n'
-# Let create a directory to hold all the cdk_nag results from ASH
-#we name the directory
+
 cd /app
+
+#
+# Create a directory to hold all the cdk_nag results from ASH
+#
 DIRECTORY="ash_cf2cdk_output"
-# Lets check if this directory already exist from previous ASH run
+# Check if this directory already exist from previous ASH run
 if [ -d "$DIRECTORY" ]; then
-  #lets delete this directory and its files and recreate it.
+  # Delete this directory and its files and recreate it.
   rm -rf $DIRECTORY
-  mkdir $DIRECTORY
-else  #lets create this directory again to capture output from this run
-  mkdir $DIRECTORY
+fi
+mkdir $DIRECTORY
+
+RC=0
+
+rm /app/cdk_report_result.txt 2>/dev/null
+touch /app/cdk_report_result.txt
+
+#
+# Uncomment the diagnostic output below to get details about
+# the  environment and node versions
+#
+
+# echo "Environment:" >>/app/cdk_report_result.txt
+# echo "Node information:" >>/app/cdk_report_result.txt
+# node --version >>/app/cdk_report_result.txt
+# echo "----------------------" >>/app/cdk_report_result.txt
+# echo "Installed NPM packages:" >>/app/cdk_report_result.txt
+# npm list -g >>/app/cdk_report_result.txt
+# echo "----------------------" >>/app/cdk_report_result.txt
+# echo "CDK information:" >>/app/cdk_report_result.txt
+# cdk --version >>/app/cdk_report_result.txt
+# echo "----------------------" >>/app/cdk_report_result.txt
+
+echo -e "\nstarting to investigate ..." >>/app/cdk_report_result.txt
+
+cfn_files=($(readlink -f $(grep -lri 'AWSTemplateFormatVersion' . --exclude-dir={cdk.out,utils,.aws-sam,ash_cf2cdk_output} --exclude=ash) 2>/dev/null))
+
+cd /utils/cfn-to-cdk/
+
+if [ "${#cfn_files[@]}" -gt 0 ]; then
+  echo "found ${#cfn_files[@]} files to scan.  Starting scans ..." >>/app/cdk_report_result.txt
+
+  for file in ${cfn_files[@]}; do
+    #echo $cfn_files
+    file1=`basename $file`
+    echo ">>>>>> begin cdk-nag result for ${file1} >>>>>>" >> /app/cdk_report_result.txt
+    #
+    # Generate the CDK application inserting the CloudFormation template
+    #
+    /usr/bin/python3 cfn_to_cdk/template_generator.py $file
+    #
+    # Use CDK to synthesize the CDK application,
+    # running CDK-NAG on the inserted CloudFormation template
+    #
+    cdk synth --quiet 2>> /app/cdk_report_result.txt
+    CRC=$?
+    echo "<<<<<< end cdk-nag result for ${file1} <<<<<<" >> /app/cdk_report_result.txt
+    mkdir -p /app/$DIRECTORY/${file1}_cdk_nag_results
+    mv cdk.out/cfn-to-cdk.template.json /app/$DIRECTORY/${file1}_cdk_nag_results/
+    mv cdk.out/AwsSolutions-*-NagReport.csv /app/$DIRECTORY/${file1}_cdk_nag_results/
+    RC=$(bumprc $RC $CRC)
+  done
+else 
+  echo "found ${#cfn_files[@]} files to scan.  Skipping scans." >>/app/cdk_report_result.txt
 fi
 
-touch /app/cdk_report_result.txt
-cfn_files=$(readlink -f $(grep -lri 'AWSTemplateFormatVersion' . --exclude-dir={cdk.out,utils,.aws-sam} --exclude=ash))
-
-
-for file in $cfn_files
-do
-  #echo $cfn_files
-  file1=`basename $file`
-  echo ">>>>>> begin cdk-nag result for ${file1} >>>>>>" >> /app/cdk_report_result.txt
-  cd /utils/cfn-to-cdk/
-  /usr/bin/python3 cfn_to_cdk/template_generator.py $file
-  /usr/bin/python3 -m pip install -U -r requirements.txt > /dev/null
-  /usr/bin/python3 /utils/cdk-addon-py.py
-  cdk synth 2>> /app/cdk_report_result.txt
-  echo "<<<<<< end cdk-nag result for ${file1} <<<<<<" >> /app/cdk_report_result.txt
-  mkdir /app/$DIRECTORY/${file1}_cdk_nag_results
-  mv cdk.out/cfn-to-cdk.template.json /app/$DIRECTORY/${file1}_cdk_nag_results/
-  mv cdk.out/AwsSolutions-*-NagReport.csv /app/$DIRECTORY/${file1}_cdk_nag_results/
-  CRC=$?
-  RC=$(bumprc $RC $CRC)
-done
-
 cd /app
+
 unset IFS
 
 exit $RC