+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-<a name="1.0.8-e-03May2023"></a>
+<a name="1.0.9-e-16May2023"></a>
+# 1.0.9-e-16May2023
+### Automated Security Helper
+
+* Changed YAML scanning (presumed CloudFormation templates) to look for CloudFormation template files explicitly, and excluding some well known folders
+added additional files that checkov knows how to scan to the list of CloudFormation templates (Dockerfiles, .gitlab-ci.yml)
+* Re-factored CDK scanning in several ways:
+    * Moved Python package install to the Dockerfile (container image build) so it's done once
+    * Removed code that doesn't do anything
+    * Added diagnostic information to report regarding the CDK version, Node version, and NPM packages installed.
+* Fixed Semgrep exit code
+
 # 1.0.8-e-03May2023
 ### Automated Security Helper